scripts: make demo_common.sh usable on its own
When looking for $root_dir, do not try to guess if the root path is
Mbed TLS or TF-PSA-Crypto one, but simply look for
"scripts/project_name.txt" file without reading it. In this way the
initial part of the script does not need "project_detection.sh".

Once the root path is found, we can easily source "project_detection.sh"
and use functions defined there.

This commit also updates "dlopen_demo.sh" so that it simply sources
"demo_common.sh" and not "project_detection.sh" (not directly at least).

Signed-off-by: Valerio Setti <vsetti@baylibre.com>